NavBar converted into component
