+++ src/components/DynamicEventsPage/DynamicEventsPage.jsx
@@ -1,15 +1,38 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { useParams } from 'react-router-dom';
+import NavBar from '../NavBar.jsx';
 
 
 const DynamicEventsPage = () => {
   const { title } = useParams(); // Get the title from the URL parameters
+  const eventData = [
+    {
+      title: 'Technical Event',
+      description: 'Experience tranquil waters reflecting the surrounding landscape.',
+      backgroundImage: 'https://picsum.photos/240/320?random=4'
+    },
+    {
+      title: 'Cultural Event',
+      description: 'Experience tranquil waters reflecting the surrounding landscape.',
+      backgroundImage: 'https://picsum.photos/240/320?random=4'
+    },
+    {
+      title: 'Sports Event',
+      description: 'Experience tranquil waters reflecting the surrounding landscape.',
+      backgroundImage: 'https://picsum.photos/240/320?random=4'
+    },
+  ]
 
+  // useEffect(()=>{
+
+  // })
   return (
     <div className="p-4">
-      <h1 className="text-3xl font-bold">{decodeURIComponent(title)}</h1> {/* Display the title */}
-      {/* Add more content related to the dynamic page here */}
-      <p>This is the dynamic content for the {decodeURIComponent(title)} event.</p>
+      <NavBar />
+      <h1 className="text-4xl font-bold">{decodeURIComponent(title)}</h1>
+      <div className="">
+
+      </div>
     </div>
   );
 }
